scraper/requirements.txt
requests
python-dotenv
python-dateutil
bs4
pandas